refactor: ltl2hoapp
We now parse the LTL expression using hoapp's own parser. This allows
to catch type errors quickly and simplifies the input language (the
user is no longer required to quote expressions by hand).
On the flip side, the range of supported LTL operators at the moment
is much smaller than what Spot has to offer. That's left for future
work.